Fixed future dates bug
