+++ client/src/constants/validation.js
@@ -205,9 +205,9 @@ export const HiredParticipantSchema = yup.object().noUnknown('Unknown field in f
   phoneNumber: yup.string().required('Phone number is required').matches(/^[0-9]{10}$/, 'Phone number must be provided as 10 digits'),
   emailAddress: yup.string().required('Email address is required').matches(/^(.+@.+\..+)?$/, 'Invalid email address'),
   origin: yup.string().required('Must indicate origin of offer'),
-  contactedDate: yup.string().required('Contacted date is required').matches(/^\d{4}\/\d{2}\/\d{2}$/, 'Invalid Date'),
-  startDate: yup.string().required('Start date is required').matches(/^\d{4}\/\d{2}\/\d{2}$/, 'Invalid Date'),
-  offerDate: yup.string().required('Offer acceptance date is required').matches(/^\d{4}\/\d{2}\/\d{2}$/, 'Invalid Date'),
+  contactedDate: yup.string().required('Date of contact is required').test('is-date', 'Not a valid date in the past', validatePastDateString),
+  startDate: yup.string().required('Start date is required').test('is-date', 'Not a valid date', validateDateString),
+  offerDate: yup.string().required('Date offer accepted  is required').test('is-date', 'Not a valid date in the past', validatePastDateString),
   nonHcapOpportunity: yup.boolean().required('Non-Hcap Opportunity is required as true or false'),
   acknowledge: yup.boolean().test('is-true', 'Must acknowledge participant acceptance', (v) => v === true),
   site: yup.number().required('Site is required'),
